--- extras/Hadrons/Modules/MContraction/A2AMeson.cc
@@ -1,7 +1,0 @@
-#include <Grid/Hadrons/Modules/MContraction/A2AMeson.hpp>
-
-using namespace Grid;
-using namespace Hadrons;
-using namespace MContraction;
-
-template class Grid::Hadrons::MContraction::TA2AMeson<FIMPL>;
