+++ extras/Hadrons/Modules/MContraction/MesonFieldGmu.cc
@@ -1,0 +1,7 @@
+#include <Grid/Hadrons/Modules/MContraction/MesonFieldGmu.hpp>
+
+using namespace Grid;
+using namespace Hadrons;
+using namespace MContraction;
+
+template class Grid::Hadrons::MContraction::TMesonFieldGmu<FIMPL>;
